+++ docs/pages/guides/_meta.js
@@ -1,10 +1,9 @@
 export default {
-  "replicating-onchain-state": "Replicating onchain state",
   "hello-world": "Hello World",
   "extending-a-world": "Extending a World",
   "adding-delegation": "Adding Delegation",
   "modules": "Writing MUD Modules",
   emojimon: "Emojimon",
   testing: "Testing",
-  "best-practices": "Best Practices",
+  "replicating-onchain-state": "Replicating onchain state",
 };