+++ docs/pages/services/indexer/_meta.js
@@ -1,7 +1,0 @@
-export default {
-  "using-indexer": "Using the Indexer",
-  "sqlite-indexer": "SQLite Indexer",
-  "postgres-event-only": "PostgreSQL for events",
-  "postgres-decoded": "PostgreSQL for data (and events)",
-  sql: "SQL API (Experimental)",
-};