+++ packages/common/src/writeContract.ts
@@ -58,11 +58,9 @@ export async function writeContract<
   const account = parseAccount(rawAccount);
   const chain = client.chain;
 
-  const blockTag = "pending";
   const nonceManager = await getNonceManager({
     client: opts.publicClient ?? client,
     address: account.address,
-    blockTag,
     queueConcurrency: opts.queueConcurrency,
   });
 
@@ -78,7 +76,8 @@ export async function writeContract<
         async () => {
           const nonce = nonceManager.nextNonce();
           const params = {
-            blockTag,
+            // viem_writeContract internally estimates gas, which we want to happen on the pending block
+            blockTag: "pending",
             ...request,
             nonce,
             ...feeRef.fees,
@@ -98,6 +97,12 @@ export async function writeContract<
               debug("got nonce error, retrying", error.message);
               return;
             }
+
+            if (String(error).includes("transaction underpriced")) {
+              debug("got transaction underpriced error, retrying", error.message);
+              return;
+            }
+
             throw error;
           },
         },