+++ packages/explorer/src/app/(explorer)/worlds/[worldAddress]/page.tsx
@@ -1,5 +1,0 @@
-import { redirect } from "next/navigation";
-
-export default async function WorldPage({ params }: { params: { worldAddress: string } }) {
-  return redirect(`/worlds/${params.worldAddress}/explorer`);
-}