+++ packages/explorer/src/app/api/world/route.ts
@@ -1,24 +1,24 @@
 import { AbiFunction, Address, Hex, createWalletClient, http, parseAbi } from "viem";
 import { getBlockNumber, getLogs } from "viem/actions";
-import { getRpcUrl } from "@latticexyz/common/foundry";
 import { helloStoreEvent } from "@latticexyz/store";
 import { helloWorldEvent } from "@latticexyz/world";
 import { getWorldAbi } from "@latticexyz/world/internal";
+import { SupportedChainIds, supportedChainsById, validateChainId } from "../../../common";
 
 export const dynamic = "force-dynamic";
 
-async function getClient() {
-  const profile = process.env.FOUNDRY_PROFILE;
-  const rpc = await getRpcUrl(profile);
+async function getClient(chainId: SupportedChainIds) {
+  const chain = supportedChainsById[chainId];
   const client = createWalletClient({
-    transport: http(rpc),
+    chain,
+    transport: http(),
   });
 
   return client;
 }
 
-async function getParameters(worldAddress: Address) {
-  const client = await getClient();
+async function getParameters(chainId: SupportedChainIds, worldAddress: Address) {
+  const client = await getClient(chainId);
   const toBlock = await getBlockNumber(client);
   const logs = await getLogs(client, {
     strict: true,
@@ -36,15 +36,17 @@ async function getParameters(worldAddress: Address) {
 
 export async function GET(req: Request) {
   const { searchParams } = new URL(req.url);
-  const worldAddress = searchParams.get("address") as Hex;
+  const worldAddress = searchParams.get("worldAddress") as Hex;
+  const chainId = Number(searchParams.get("chainId"));
+  validateChainId(chainId);
 
   if (!worldAddress) {
     return Response.json({ error: "address is required" }, { status: 400 });
   }
 
   try {
-    const client = await getClient();
-    const { fromBlock, toBlock, isWorldDeployed } = await getParameters(worldAddress);
+    const client = await getClient(chainId);
+    const { fromBlock, toBlock, isWorldDeployed } = await getParameters(chainId, worldAddress);
     const worldAbiResponse = await getWorldAbi({
       client,
       worldAddress,