+++ packages/explorer/src/queries/useAbiQuery.ts
@@ -1,11 +1,11 @@
 import { useParams } from "next/navigation";
 import { AbiFunction, Hex } from "viem";
 import { UseQueryResult, useQuery } from "@tanstack/react-query";
+import { useChain } from "../hooks/useChain";
 
-export async function getAbi(worldAddress: Hex) {
-  const res = await fetch(`/api/world?${new URLSearchParams({ address: worldAddress })}`);
+export async function getAbi(chainId: number, worldAddress: Hex) {
+  const res = await fetch(`/api/world?${new URLSearchParams({ chainId: String(chainId), worldAddress })}`);
   const data = await res.json();
-
   if (!res.ok) {
     throw new Error(data.error);
   }
@@ -20,9 +20,11 @@ type AbiQueryResult = {
 
 export const useAbiQuery = (): UseQueryResult<AbiQueryResult> => {
   const { worldAddress } = useParams();
+  const { id: chainId } = useChain();
+
   return useQuery({
-    queryKey: ["abi", worldAddress],
-    queryFn: () => getAbi(worldAddress as Hex),
+    queryKey: ["abi", chainId, worldAddress],
+    queryFn: () => getAbi(chainId, worldAddress as Hex),
     select: (data) => {
       return {
         abi: data.abi || [],