+++ packages/store/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 2.2.3
+
+### Patch Changes
+
+- @latticexyz/common@2.2.3
+- @latticexyz/config@2.2.3
+- @latticexyz/protocol-parser@2.2.3
+- @latticexyz/schema-type@2.2.3
+
 ## 2.2.2
 
 ### Patch Changes